Guardar lapiz no exede el ancho ni el largo, pasa goma
