+++ src/components/WhiteBoardDashBoard/components/WhiteBoard/utils/cursorIcons.js
@@ -1,0 +1,3 @@
+// cursorIcons.js
+export   const pencilCursor ='url("data:image/svg+xml,%3Csvg xmlns=%27http://www.w3.org/2000/svg%27 viewBox=%270%200%2024%2024%27 fill=%27none%27 stroke=%27black%27 stroke-width=%272%27 stroke-linecap=%27round%27 stroke-linejoin=%27round%27 class=%27feather feather-edit-2%27%3E%3Cpath d=%27M17 3L21 7L7 21H3V17L17 3Z%27/%3E%3C/svg%3E"), auto';
+export const eraserCursor = 'url("data:image/svg+xml,%3Csvg xmlns=%27http://www.w3.org/2000/svg%27 viewBox=%270%200%2024%2024%27 fill=%27none%27 stroke=%27black%27 stroke-width=%272%27 stroke-linecap=%27round%27 stroke-linejoin=%27round%27 class=%27feather feather-eraser%27%3E%3Cpath d=%27M20.17 7.83a4 4 0 0 0-5.66 0l-8 8a4 4 0 0 0 0 5.66l1.17 1.17a4 4 0 0 0 5.66 0l8-8a4 4 0 0 0 0-5.66z%27/%3E%3Cpath d=%27M12 12l-4 4%27/%3E%3Cpath d=%27M4 20h16%27/%3E%3C/svg%3E"), auto';
