+++ src/components/WhiteBoardDashBoard/components/WhiteBoard/utils/drawingUtils.js
