--- src/components/Whiteboard/ErrorBoundary.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError(error) {
-    return { hasError: true };
-  }
-
-  componentDidCatch(error, errorInfo) {
-    console.error("Error:", error, errorInfo);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return <h1>Algo salió mal.</h1>;
-    }
-    return this.props.children; 
-  }
-}
-
-export default ErrorBoundary;
